+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * Parse `<options>` that don't have value attribute (#85).
 
+* Remove all remaining uses of `html()` in favor of `read_html()` (@jimhester,
+  #113).
+
 # rvest 0.3.0
 
 * rvest has been rewritten to take advantage of the new xml2 package. xml2